Allow Path to concatenate with integers. Fix #146
As an `int` can only represent a single index, we can use handle `P + int` the same way we handle `P[int]`:
`Path(...) + 1 == Path(...)[1]`
`1 + Path(...) == Path()[1][Path(...)]`

An alternate implementation would be to convert `1` to `"[1]"` and let it parse as a string.

Corresponding tests will be updated in the folowing commits.

The case for `Path(int)` is independent of this and is not addressed in this commit.